fundamental issue.
I can format by date. however, due to the nature of the Account class - it will only ever be simple addition/subtraction - dates will make no difference to the order -> it seems that it will just do the calculations in the order they come in - which might be ok... but in terms of bullet proof testing it fails.