complete
as far as I can see this covers the need for this code. I'm not happy with the date sorting issue but functionally it seems to work -> this may not be an issue in the grand scheme of things but frustrating none the less.

not every line of this code was handwritten by me and a lot was sampled and then modified - though I believe I understand everything that is used in the core functionality.